Begin Hanami sync (#73)
Split the sync into separate jobs (backed by a reusable workflow) and start syncing the Hanami gems. For now, start with hanami/hanami and a handful of files only.

This does include `.github/workflows/ci.yml`, which we expand to allow the repos to declare (a) whether they want node (we'll default to a hard-coded version) and (b) any matrix dimensions they have. To use this, I added the following to hanami/hanami's `repo-sync.yml`:

```yaml
ci:
  matrix:
    rack: ["~> 2.0", "~> 3.0"]
  node: true
```