Post comments via dedicated workflow (#78)
This allows for comments to be posted even for PRs coming from untrusted users and forks, where our main workflows have a read-only `GITHUB_TOKEN`. 

One option for solving this would be for our main workflows to use `on: pull_request_target`, which provides a read-write `GITHUB_TOKEN`. This would allow the comments to post, but would mean greater security risk because the whole workflow would be running with those elevated permissions for code changes from an unknown party.

So instead we take a two-phased approach for posting comments:

1. The original workflows run with their standard read-only permissions for fork PRs. Instead of attempting to post PR comments directly, they generate artifacts representing those comments.
2. We then run a second workflow with the `on: workflow_run` trigger, which has full read-write access, and is separate from the untrusted code. This downloads all comment artifacts and posts comments from their data.

To ensure this approach is maintainable and the workflow YAML concise, provide this functionality via two reusable actions hosted within this repo: `pr-comment-artifact` and `pr-comments-from-artifacts`.

This is based on the approach outlined in https://securitylab.github.com/resources/github-actions-preventing-pwn-requests/.